Move exampleoc out of internal for now to unblock
