+++ exampleoc/README.md
@@ -1,0 +1,1 @@
+This package MAY change without warning, and if possible SHOULD NOT be imported into your application.
