+++ exampleoc/gen.go
