+++ exampleoc/modelb/modelb-0.go
@@ -16,18 +16,18 @@
 Package modelb is a generated package which contains definitions
 of structs which generate gNMI paths for a YANG schema.
 
-This package was generated by ygnmi version: (devel): (ygot: v0.29.8-0.20230731012118-cbc5a1ba9160)
+This package was generated by ygnmi version: (devel): (ygot: (devel))
 using the following YANG input files:
-  - ../../pathgen/testdata/yang/openconfig-simple.yang
-  - ../../pathgen/testdata/yang/openconfig-withlistval.yang
-  - ../../pathgen/testdata/yang/openconfig-nested.yang
+  - ../pathgen/testdata/yang/openconfig-simple.yang
+  - ../pathgen/testdata/yang/openconfig-withlistval.yang
+  - ../pathgen/testdata/yang/openconfig-nested.yang
 
 Imported modules were sourced from:
 */
 package modelb
 
 import (
-	oc "github.com/openconfig/ygnmi/internal/exampleoc"
+	oc "github.com/openconfig/ygnmi/exampleoc"
 	"github.com/openconfig/ygnmi/ygnmi"
 	"github.com/openconfig/ygot/ygot"
 	"github.com/openconfig/ygot/ytypes"
