+++ exampleoc/schema.go
@@ -18,11 +18,11 @@ of structs which represent a YANG schema. The generated schema can be
 compressed by a series of transformations (compression was true
 in this case).
 
-This package was generated by ygnmi version: (devel): (ygot: v0.29.8-0.20230731012118-cbc5a1ba9160)
+This package was generated by ygnmi version: (devel): (ygot: (devel))
 using the following YANG input files:
-  - ../../pathgen/testdata/yang/openconfig-simple.yang
-  - ../../pathgen/testdata/yang/openconfig-withlistval.yang
-  - ../../pathgen/testdata/yang/openconfig-nested.yang
+  - ../pathgen/testdata/yang/openconfig-simple.yang
+  - ../pathgen/testdata/yang/openconfig-withlistval.yang
+  - ../pathgen/testdata/yang/openconfig-nested.yang
 
 Imported modules were sourced from:
 */
