+++ internal/uexampleoc/enum_map.go
@@ -18,7 +18,7 @@ of structs which represent a YANG schema. The generated schema can be
 compressed by a series of transformations (compression was false
 in this case).
 
-This package was generated by ygnmi version: (devel): (ygot: v0.29.8-0.20230731012118-cbc5a1ba9160)
+This package was generated by ygnmi version: (devel): (ygot: (devel))
 using the following YANG input files:
   - ../../pathgen/testdata/yang/openconfig-simple.yang
   - ../../pathgen/testdata/yang/openconfig-withlistval.yang
