+++ internal/uexampleoc/nested/nested-0.go
@@ -16,7 +16,7 @@
 Package nested is a generated package which contains definitions
 of structs which generate gNMI paths for a YANG schema.
 
-This package was generated by ygnmi version: (devel): (ygot: v0.29.8-0.20230731012118-cbc5a1ba9160)
+This package was generated by ygnmi version: (devel): (ygot: (devel))
 using the following YANG input files:
   - ../../pathgen/testdata/yang/openconfig-simple.yang
   - ../../pathgen/testdata/yang/openconfig-withlistval.yang
